export track bounds utils
