+++ src/map/index.js
@@ -1,3 +1,4 @@
 export { default } from './map'
 export { fitToBounds, targetMapVessel } from './store'
+export { getTrackBounds, getTrackTimeBounds } from './utils/getTrackBounds'
 export { default as AVAILABLE_BASEMAPS } from './basemaps'
